remove prefix, divide up device classes
